kernel: add atomic types, add proc.h
Moves task_struct definition to proc.h and add atomic types.
