+++ include/nyx/proc.h
@@ -1,0 +1,33 @@
+#ifndef _NYX_PROC_H
+#define _NYX_PROC_H
+
+#include <mm/mm_types.h>
+#include <uapi/posix_types.h>
+
+#include <asi/proc.h>
+
+#define TASK_NAME_LEN 32
+
+enum task_state {
+    TASK_NEW,
+    SLEEPING,
+    RUNNABLE,
+    RUNNING,
+    ZOMBIE,
+};
+
+struct task_struct {
+    struct list_head list;
+
+    enum task_state     state;
+    struct proc_context context;
+    struct vas_struct  *mm;
+    struct vas_struct  *active_mm;
+    pid_t               pid;
+    void               *stack;
+
+    struct list_head task_list;
+    char             name[TASK_NAME_LEN];
+};
+
+#endif
