+++ include/nyx/sched.h
@@ -1,37 +1,10 @@
 #ifndef _NYX_SCHED_H
 #define _NYX_SCHED_H
 
-#include <mm/mm_types.h>
 #include <nyx/compiler.h>
 #include <nyx/list.h>
+#include <nyx/proc.h>
 #include <nyx/types.h>
-#include <uapi/posix_types.h>
-
-#include <asi/proc.h>
-
-#define TASK_NAME_LEN 32
-
-enum task_state {
-    TASK_NEW,
-    SLEEPING,
-    RUNNABLE,
-    RUNNING,
-    ZOMBIE,
-};
-
-struct task_struct {
-    struct list_head list;
-
-    enum task_state     state;
-    struct proc_context context;
-    struct mm_struct   *mm;
-    struct mm_struct   *active_mm;
-    pid_t               pid;
-    void               *stack;
-
-    struct list_head task_list;
-    char             name[TASK_NAME_LEN];
-};
 
 struct task_struct *get_current_task();
 void                schedule();
